+++ src/facet/__init__.py
@@ -1,7 +1,9 @@
 import importlib.metadata
 
-from facet.asyncio import ServiceMixin
+from facet.asyncio import AsyncioServiceMixin
+from facet.asyncio import AsyncioServiceMixin as ServiceMixin
+from facet.blocking import BlockingServiceMixin
 
-__all__ = ("ServiceMixin", "__version__", "version")
+__all__ = ("AsyncioServiceMixin", "ServiceMixin", "BlockingServiceMixin", "__version__", "version")
 __version__ = importlib.metadata.version(__package__)
 version = tuple(map(int, __version__.split(".")))